requirements.txt
djangorestframework>=3.14
django>=4.2
PyMySQL>=1.0
mysqlclient>=2.2  
requests>=2.28
openpyxl>=3.1.2  
django-tables2>=2.6  
django-filter>=23.3  
django-crispy-forms>=2.0  
crispy-bootstrap5>=0.7  
